+++ node/directives/helper.ts
@@ -2,12 +2,12 @@ import { isUserPartOfBuyerOrg } from '../resolvers/Queries/Users'
 
 export const validateAdminToken = async (
   context: Context,
-  adminUserAuthToken: string
+  adminUserAuthToken: string,
+  metricFields: any
 ): Promise<{
   hasAdminToken: boolean
   hasValidAdminToken: boolean
   hasCurrentValidAdminToken: boolean
-  hasValidAdminTokenFromStore: boolean
 }> => {
   const {
     clients: { identity, lm },
@@ -19,8 +19,6 @@ export const validateAdminToken = async (
   let hasValidAdminToken = false
   // this is used to check if the token is valid by current standards
   let hasCurrentValidAdminToken = false
-  // this is used to check if the token is valid and from this store
-  let hasValidAdminTokenFromStore = false
 
   if (hasAdminToken) {
     try {
@@ -32,17 +30,22 @@ export const validateAdminToken = async (
       // in the future we should remove this line
       hasCurrentValidAdminToken = true
 
-      if (authUser?.audience === 'admin') {
-        hasValidAdminToken = await lm.getUserAdminPermissions(
-          authUser.account,
-          authUser.id
-        )
+      if (
+        authUser?.audience === 'admin' &&
+        authUser &&
+        authUser.account !== account
+      ) {
+        logger.warn({
+          message: 'validateAdminToken: Token from another account',
+          ...metricFields,
+          authUserAccount: authUser.account,
+          account,
+          user: authUser.user,
+        })
       }
 
-      // check if the token is from this store. Currently used for metrics
-      // in future we should merge this with the previous check
       if (authUser?.audience === 'admin' && authUser?.account === account) {
-        hasValidAdminTokenFromStore = await lm.getUserAdminPermissions(
+        hasValidAdminToken = await lm.getUserAdminPermissions(
           account,
           authUser.id
         )
@@ -60,16 +63,15 @@ export const validateAdminToken = async (
     hasAdminToken,
     hasValidAdminToken,
     hasCurrentValidAdminToken,
-    hasValidAdminTokenFromStore,
   }
 }
 
 export const validateApiToken = async (
-  context: Context
+  context: Context,
+  metricFields: any
 ): Promise<{
   hasApiToken: boolean
   hasValidApiToken: boolean
-  hasValidApiTokenFromStore: boolean
 }> => {
   const {
     clients: { identity, lm },
@@ -81,7 +83,6 @@ export const validateApiToken = async (
   const appKey = context?.headers['vtex-api-appkey'] as string
   const hasApiToken = !!(apiToken?.length && appKey?.length)
   let hasValidApiToken = false
-  let hasValidApiTokenFromStore = false
 
   if (hasApiToken) {
     try {
@@ -94,14 +95,22 @@ export const validateApiToken = async (
         token,
       })
 
-      if (authUser?.audience === 'admin') {
-        hasValidApiToken = true
+      if (
+        authUser?.audience === 'admin' &&
+        authUser &&
+        authUser.account !== account
+      ) {
+        logger.warn({
+          message: 'validateApiToken: Token from another account',
+          ...metricFields,
+          authUserAccount: authUser.account,
+          account,
+          user: authUser.user,
+        })
       }
 
-      // check if the token is from this store. Currently used for metrics
-      // in future we should merge this with the previous check
       if (authUser?.audience === 'admin' && authUser?.account === account) {
-        hasValidApiTokenFromStore = await lm.getUserAdminPermissions(
+        hasValidApiToken = await lm.getUserAdminPermissions(
           account,
           authUser.id
         )
@@ -115,7 +124,7 @@ export const validateApiToken = async (
     }
   }
 
-  return { hasApiToken, hasValidApiToken, hasValidApiTokenFromStore }
+  return { hasApiToken, hasValidApiToken }
 }
 
 export const validateStoreToken = async (
@@ -168,7 +177,8 @@ export const validateStoreToken = async (
 }
 
 export const validateAdminTokenOnHeader = async (
-  context: Context
+  context: Context,
+  metricFields: any
 ): Promise<{
   hasAdminTokenOnHeader: boolean
   hasValidAdminTokenOnHeader: boolean
@@ -186,7 +196,7 @@ export const validateAdminTokenOnHeader = async (
   }
 
   const { hasAdminToken, hasCurrentValidAdminToken, hasValidAdminToken } =
-    await validateAdminToken(context, adminUserAuthToken)
+    await validateAdminToken(context, adminUserAuthToken, metricFields)
 
   return {
     hasAdminTokenOnHeader: hasAdminToken,