+++ node/metrics/auth.ts
@@ -18,8 +18,6 @@ export interface AuthAuditMetric {
   hasValidApiToken?: boolean
   hasAdminTokenOnHeader?: boolean
   hasValidAdminTokenOnHeader?: boolean
-  hasValidAdminTokenFromStore?: boolean
-  hasValidApiTokenFromStore?: boolean
 }
 
 export class AuthMetric implements Metric {