+++ CHANGELOG.md
@@ -11,6 +11,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - addUser function to not accept invalid cost center
 
+## [1.43.3] - 2024-07-31
+
+### Changed
+- Changed the token validation directive of some operations
+
+## [1.43.2] - 2024-07-29
+
+### Added
+- Add enforcement of new validation for admin and api tokens
+- Add more details to admin and api token validation metric
+
 ## [1.43.1] - 2024-07-24
 
 ### Changed